+++ lib/profile/WP_Auth0_Profile_Change_Password.php
@@ -1,4 +1,11 @@
 <?php
+/**
+ * Contains class WP_Auth0_Profile_Change_Password.
+ *
+ * @package WP-Auth0
+ *
+ * @since 3.8.0
+ */
 
 /**
  * Class WP_Auth0_Profile_Change_Password.
@@ -27,8 +34,15 @@ class WP_Auth0_Profile_Change_Password {
 	 * @codeCoverageIgnore - Tested in TestProfileChangePassword::testInitHooks()
 	 */
 	public function init() {
+
+		// Used during profile update in wp-admin.
 		add_action( 'user_profile_update_errors', array( $this, 'validate_new_password' ), 10, 2 );
+
+		// Used during password reset on wp-login.php
 		add_action( 'validate_password_reset', array( $this, 'validate_new_password' ), 10, 2 );
+
+		// Used during WooCommerce edit account save.
+		add_action( 'woocommerce_save_account_details_errors', array( $this, 'validate_new_password' ), 10, 2 );
 	}
 
 	/**
@@ -36,22 +50,27 @@ class WP_Auth0_Profile_Change_Password {
 	 * Hooked to: user_profile_update_errors, validate_password_reset
 	 * IMPORTANT: Internal callback use only, do not call this function directly!
 	 *
-	 * @param WP_Error        $errors - WP_Error object to use if validation fails.
-	 * @param boolean|WP_User $user - Boolean update or WP_User instance, depending on action.
+	 * @param WP_Error         $errors - WP_Error object to use if validation fails.
+	 * @param boolean|stdClass $user - Boolean update or WP_User instance, depending on action.
 	 *
 	 * @return boolean
 	 */
 	public function validate_new_password( $errors, $user ) {
 
 		// Exit if we're not changing the password.
-		if ( empty( $_POST['pass1'] ) ) {
+		// The pass1 key is for core WP, password_1 is WooCommerce.
+		if ( empty( $_POST['pass1'] ) && empty( $_POST['password_1'] ) ) {
 			return false;
 		}
-		$new_password = $_POST['pass1'];
+
+		$field_name   = ! empty( $_POST['pass1'] ) ? 'pass1' : 'password_1';
+		$new_password = $_POST[ $field_name ];
 
 		if ( isset( $_POST['user_id'] ) ) {
+			// Input field from user edit or profile update.
 			$wp_user_id = absint( $_POST['user_id'] );
-		} elseif ( is_object( $user ) && $user instanceof WP_User ) {
+		} elseif ( is_object( $user ) && ! empty( $user->ID ) ) {
+			// User object passed in from an action.
 			$wp_user_id = absint( $user->ID );
 		} else {
 			return false;
@@ -83,7 +102,7 @@ class WP_Auth0_Profile_Change_Password {
 
 		// Add an error message to appear at the top of the page.
 		$error_msg = is_string( $result ) ? $result : __( 'Password could not be updated.', 'wp-auth0' );
-		$errors->add( 'auth0_password', $error_msg, array( 'form-field' => 'pass1' ) );
+		$errors->add( 'auth0_password', $error_msg, array( 'form-field' => $field_name ) );
 		return false;
 	}
 }